ability to set ImageInfo adjoin flag
This flag changes the behavior of `load_image' and `write_image' when
the image has multiple frames. For instance, on `write_image` it can be
used to create multiple images, one for each frame.

The image_dump image has been rewritten to use `WriteImages` instead of
`WriteImage`. This functions does the same thing as `WriteImage` but can
use the `adjoin` flag.