new `flag` function and new ! functions
The `flag` function allows you to set/unset a boolean flag on the
image. Currently only `adjoin` is supported. Additionaally, all
functions now have the bang version so it is easier to use them with the
pipe operator:

```
ExMagick.image!
|> ExMagick.load_image!
|> ExMagick.flag!
|> ExMagick.dump_image!
```

The bang version will raise an exception on error but you can always
refer to the the non-bang version of it that returns a tuple.