bugfix: memory leak in ImageToBlob
image_dump!/1 is leaking memory because we are not deallocating memory
after `ImageToBlob`. the following script should exhaust the machine's
memory after a short while:

    ```
    defmodule Loop do
      def loop() do
        ExMagick.init!()
        |> ExMagick.image_load!("/tmp/foobar.tiff") # 16MB file
        |> ExMagick.image_dump!()
        loop()
      end
    end
    Loop.loop()
    ```

the solution is to copy the blob into a new new managed binary. this
might impact performance a bit depending on the size of the image but
as far as we know this can't be helped. it seems there is no way to
attach a destructor to a nif term other than `enif_open_resource_type`
which can't occur outside of `load` or `upgrade` making this extra
`memcpy` unavoidable.